requirements.txt
tensorflow==2.15.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.23.5
matplotlib==3.7.1
pillow==9.5.0
flask==2.2.5
gunicorn==21.2.0
werkzeug==3.0.1
keras==2.15.0
